UndoRedoStack and UndoRedoManager tests "complete"

+++ UndoRedoFramework/Test/Calculator.h
@@ -9,18 +9,18 @@ public:
 	Calculator() : result(0){};
 	~Calculator(){};
 
-	void plus(double summand){
+	void plus(int summand){
 		std::cout << "Calculator.plus: " << summand << std::endl;
 		result += summand;
 	}
-	void minus(double subtrahent){
+	void minus(int subtrahent){
 		std::cout << "Calculator.minus: " << subtrahent << std::endl;
 		result -= subtrahent;
 	}
-	double getResult(){ return result; }
+	int getResult(){ return result; }
 	void clear() { result = 0; }
 private:
-	double result;
+	int result;
 };
